+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.3] - 2025-04-15
+
+### Added
+- Changes in install/README.md
+- Changes in install/fedora.sh
+- Changes in install/ubuntu.sh
+- Changes in install/windows.ps1
+
 ## [0.1.2] - 2025-04-15
 
 ### Added